Update specs to match reality
Here we simply update the tests to match what the code is doing.
Previously, only the tests made reference to content being an array.  I
think this is a use case we could suppoort, but it isn't one currently.
To avoid a sitaution of tail-wags-dog, here we just update the tests and
validate the input.