Change default interface value
In order to follow the unbound default, here we cease to require that an
interface is passed to the module.  Instead, we allow this parameter to be
optional.  When set, this parameter behaves just as it did before.  When the
user does not provide this value, the defaults are no longer passed in, and thus
absent from the configuration file on disk resulting in the default values
coming from unbound.conf(5) which is to listen on localhost.