update minimum unbound version
This PR updates the minimum supported version of unbund.  this allows
us to be a bit more generous when writing the first unbound.conf file
and fixes #286

I have picked unbound 1.6.6 as the minimum supported version.  this
seems to be the version available in centos 7 which i assume is the
oldest version we support based on metatdata.json