Require Unbound 1.6.6 or newer
This updates the minimum supported version of unbound. This allows
us to be a bit more generous when writing the first unbound.conf file
and fixes #286.

I have picked unbound 1.6.6 as the minimum supported version. This
seems to be the version available in CentOS 7 which I assume is the
oldest version we support based on metatdata.json.