Introduce unbound_version parameter
The version of the installed unbound instance. Defaults to the fact, but
you can overwrite it. this reduces the initial puppet runs from two to
one (because currently during the first run, the version is an empty
string and it defaults to `0a`. The module hides unbound.conf options
that are introduced in later versions.